backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==3.14.0
django-filter==24.2
djangorestframework==3.15.1
gunicorn==22.0.0
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.0
typing_extensions==4.12.2
whitenoise==6.5.0
